dex: Swap alt fee tokens for native fee token (#4643)
## Describe your changes

Moves fee handling out of the Transaction's ActionHandler and into
the component, which accumulates fees (in any token), and adds hooks to
the DEX
thatadd chain-submitted swaps for any non-native fee tokens into the
native
token. These are then accumulated back into the fee component.

The base fee and tip are tracked separately through this whole process,
in
order to support the intended behaviour (the base fee is burned to
account for
the _chain_'s resource use, the tip is sent to the proposer to encourage
them
to include transactions in fee priority order). However, tip handling is
not
currently implemented, so both are burned. Later, the
funding/distribution
components should extract the tip and send it to the proposer's funding
streams. However, until a robust fee market develops, this form of
proposer
incentivization can be deferred. The accounting will be there for it
when it
arises.

## Issue ticket number and link

#4328 

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > consensus-breaking but not state-breaking, no migrations necessary

---------

Co-authored-by: Erwan Or <erwanor@penumbralabs.xyz>