docs: tutorial for running hardware node (#4678)
## Describe your changes
   
Documents the setup process for renting a hardware box (via Hetzner) and
configuring a fullnode on it. Necessarily overlaps with the regular ol'
install docs for pd, but tried to be explicit in the tutorial for
maximum comprehension.

Making that simplified change, I looked through the old deployment/ci
configs and culled a lot that were gathering dust. Now the repo
concisely recommends using the systemd config, or links out to other
community-maintained options for configuring nodes.

## Issue ticket number and link
None.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> no application code changes, just improving docs and cleaning things
up

---------

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>