pd: rename "testnet" subcommand to "network" (#4684)
## Describe your changes
Mostly a cosmetic change, since it's a rename. Requires lots of
carry-over throughout the docs. Kept the internal APIs the same for the
most part. This is a breaking change from the context of node operators,
though, because now pd will look in a different location for info.

## Issue ticket number and link
N/A

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> This is a breaking change for the CLI, for sure, although I tried hard
not to touch critical app logic. Careful attention should be paid to
e.g. `pd network generate` and make sure the functionality remains the
same.

---------

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>