Drive `pindexer` forward with staking example (#4707)
## Describe your changes

This PR drives the `pindexer` development forward with worked examples
of staking-related data -- using https://penumbra.today as an example
use case. With these changes, it should be possible for all of the data
on that site to be rendered directly out of Postgres, without any RPC
calls.

We should figure out how to test this -- it should be amenable to
testing since we have a good source of event data (mainnet) and
expectations about what it should look like -- but this is less critical
at this instant.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > Only changes to code outside `pindexer` are convenience methods.