docs: add hermes setup steps to guide
Ported the contents from the old wiki page (since cleaned up) on configuring
Hermes to the proper dev guide, beside the other node operator docs.
Most of the advice is "use this fork" and "add these specific" options,
but at least now there's a single URL to collect that info. We can
revise once the relevant Penumbra-compatibility changes in the fork are
upstreamed.