feat(dev): add nix-based dev-env (#4726)
Trying to make it easier to spin up a fullnode on localhost. The team
has been focused on writing tooling, like pindexer, and the manual steps
to munge CometBFT configs and setup a localhost postgres database are
not awesome. In addition to pd/cometbft spin-up, the dev-env includes:

  * metrics
  * postgres
  * pindexer

Not yet including grpcui, because the latest published version doesn't
support the new server reflection API. Will tack that on in a dedicated
follow-up.

The dev-env is strongly opinionated and uses nix [0] to manage
cross-platform, polyglot dependencies. There's other great work in our
ecosystem, e.g. [1], and this early push tries to be consistent with
that.

Updates the relevant first-run docs. Much more work should be done on
the developer-focused content in the guide overall, including an
overhaul of discussing RPC endpoints and web-relevant resources, but
those changes can land in the future. This PR aims to establish a basis
for a sane, supported dev-env, providing a pleasant first-run experience
for newcomers (and existing devs too).

[0] https://nixos.org
[1] https://github.com/informalsystems/cosmos.nix