docs: add hermes setup steps to guide (#4732)
## Describe your changes
Ported the contents from the old wiki page (since cleaned up) on
configuring Hermes to the proper dev guide, beside the other node
operator docs. Most of the advice is "use this fork" and "add these
specific" options, but at least now there's a single URL to collect that
info. We can revise once the relevant Penumbra-compatibility changes in
the fork are upstreamed.

## Issue ticket number and link
N/A
## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > docs-only

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>