(backport) feat: add timestamp to all event protos (#4722) (#4734)
this makes a non-consensus-breaking change, adding a timestamp field to
the EventBlockRoot proto.
This will be used to keep track of the timestamp of the block for use in
pindexer.

I tested this by running a local devnet with the changes to core. Pd
runs fine and the timestamps are correctly propagated to the raw events
db.

prerequisite for https://github.com/penumbra-zone/penumbra/issues/4675

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

- only adds a field
- adds this field to an Event proto

## Describe your changes

## Issue ticket number and link

## Checklist before requesting a review

- [ ] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > REPLACE THIS TEXT WITH RATIONALE (CAN BE BRIEF)

---------

Co-authored-by: Josef <vacekj@outlook.com>
Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>