Make pindexer more easily embeddable as a library (#4712)
## Describe your changes

We don't expose cometindex as a library, but we do pindexer, and this
adds a few tweaks to make pindexer possible to run inside an external
piece of Rust code. This is useful, because it allows people to colocate
their indexer and their block explorer, for example.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > Doesn't touch the core protocol.