ci: move guide docs to separate repo (#4751)
## Describe your changes
The Penumbra Guide has been moved to its own repo [0], and ported from
mdbook [1] to nextra [2] in the process. As a follow-up, I'm removing
the deprecated mdbook-based guide from the protocol repo, as well as the
corresponding CI workflows.

[0] https://github.com/penumbra-zone/guide
[1] https://rust-lang.github.io/mdBook
[2] https://nextra.site

## Issue ticket number and link
N/A

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > docs and CI changes only, no changes to application code.

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>