Bug: correctly handle denoms with `/` in pcli tx lp
The parsing looked for the first occurrence of the slash for

nX@nY/fee

but if X or Y contains a slash, this will be incorrect.

This fixes this somewhat, by looking for the last slash, and only
considering that a fee if it ends with bps.

This will not correctly handle a denom containing both a slash and
ending with bps, but beyond that we'll need to change the format
probably.