events: record `EventPositionClose` when closing positions manually (#4779)
There are currently two ways for a position to get closed:

  1. automatically by the DEX, for an auto-close position,
2. at the end of the block, after being *queued* for closure in the in
the action handler.

For 2., we weren't emitting the event.
This commit changes things so that we do, in fact, emit the event in
this case, for a position that has not yet been closed.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > Only adds new events.

(cherry picked from commit 6981aec13519d352be642f690350b5227657de4f)