feat(pd)!: migration for ibc noble fixes
This commit adds a pd migration, intended to permit upgrade
coordination, so that breaking changes related to ibc withdrawal
handling can be made to pd safely. The migration doesn't alter chain
state: it simply flips the halt bit off, and permits the chain to
resume, after folks have upgraded to a new version of pd.
This approach assumes that an upgrade proposal was submitted and passed,
in order to halt the chain in a coordinated upgrade.