feat(pd)!: stateful migration for IBC noble fixes
This adds a state change to the migration, in order to replace all of
the packets which had error acknowledgements, so that, post-migration,
the acks can be replayed with the correct logic for error acks, the
packets having been replaced in the state.

This also changes the IBC component visibility of a trait to reuse the
logic for inserting packet commitments.