pcli: add un-advertised support for importing Prax registry data (#4801)
This makes a minimal set of changes to allow `pcli` to make use of the
Prax registry. In order to not pre-empt more comprehensive planning
about how such an integration should work, this PR just changes `pcli`
so that if the registry is placed in a `registry.json` file in the
`pcli` home directory, it will be imported into the view database on
startup.

(cherry picked from commit 5ed9813804af18066a22acfb2bd850e01ece3f46)