docs: add issue template for point-release
This is the issue format I've been using for e.g. v0.80.3 (#4829),
so I'm adding to the repo, to make it easier to find and reuse.
In the process, I'm cleaning out the old "release testnet" checklist,
because that's not a process we follow regularly anymore: we don't have
regular or frequent chain resets.