feat(pcli): add balance migration command
Adds a new subcommand to pcli for moving the entirety of a wallet's
assets to a new wallet, as defined by an FVK. Stores both FVKs in the
memo of the transaction. Leaves zero notes in the source wallet,
and the dest wallet gets everything, minus whatever gas price was paid.

The full-send functionality is achieved by using the change address
of the planner, setting it to the dest FVK's 0 address.
This approach avoids the issue of dealing with fees outside the planner.
It does this by eschewing trying to figure out how much to output,
and instead focusing on what to spend (everything).
Then, the change address is set to the desired destination, allowing the
planner to automatically move the funds as needed, minus fees.

Ensures that the FVK is read from stdin, so that it's never placed on
the command line.

Co-authored-by: aubrey <aubrey@penumbra.zone>
Co-authored-by: Lucas Meier <lucas@cronokirby.com>
Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>