pcli migrate balance: replicate account structure
This changes the migrate balance command to replicate the balances of
each asset per account, as required.
Previously, the command would move all the values into the 0 account.

This will pull fees from the account with the most amount of the staking
/ fee token.
This differs slightly from iterating through account numbers starting
from 0, but is much much simpler.