Basic ICS23 Transfer Test (#4850)
## Describe your changes

Replaces the existing IBC handshake test with a full ICS23 transfer
test. Only the "happy path" (no timeouts or any other error cases) is
tested in this initial spike.

Also introduces some improvements to the underlying `MockClient` to
support spent note tracking.

The transfer implementation in the `MockRelayer` is currently hardcoded
to transfer 50% of the first chain's client's first note to the second
chain; this can be fleshed out further in response to future test
requirements.

## Issue ticket number and link

Closes #4846

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > tests only