ci: add cargo-release to deps
Adds two more packages to the nix dev again:

  * cargo-release
  * buf

Both of those tools are critical during release-engineering,
so I'm adding them to provide a deterministic way to generate the
necessary build inputs.

In this commit the nix env switches to the recently published stable
release for 2024-11. That stable release pulls in `buf` version 1.47.0,
which is sufficiently recent to satisfy the "at least 1.32" stipulation
we've had in the docs [0] for a while now.

[0] https://guide.penumbra.zone/dev/protobuf#installing-buf