pindexer: implement batch processing API (#4913)
This restructures the AppView interface to allow processing events in a
batch. Previously, app views had to index one event at a time. This PR
changes things so that app views get a batch of several blocks worth of
events, with a guarantee to have all of the events in any block in the
batch.

## Making App Views Easier to Write
By having access to all the events in a block, app views are more
ergonomic to write. For example, the dex explorer app view wants to know
the time of the candlestick events it processes, but to do this, it
needs to wait for the block root event later in the block, which
provides this timestamp. Currently, because we don't have access to any
context, we need to manually implement a queuing system in the database,
which is very annoying, and a performance hit.

## Making App Views More Performant
We can make app views more performant by processing both an entire
block, and multiple blocks, since:
- we don't need to write an update more than once per block to the
database
- we may be able to write updates less frequently, depending on the app
view (e.g. when we need only the current value)
- we can keep transient state in memory, instead of on the database,
reducing writes and reads in all cases

## Additional Performance Improvements
Now the app views are run in parallel, which provides additional
improvements when syncing up.

## Testing
Pindexer should work as usual, after wiping the database.

# Checklist

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > pindexer only