Don't touch uninitialized storage for version safeguard
This prevents an edge case where PD would crash if starting before the
very first genesis. Not touching storage in that case will prevent nodes
running continuously from genesis from benefitting from the safeguard,
but an upgrade has already happened on mainnet, and so we don't care
about not having the safeguard in this case.