migrate: informative msg on re-attempted migration
If an operator runs `pd migrate` and then runs it again, the error
message should accurately describe the situation that the local state is
already migrated, and state that pd is refusing to proceed as
instructed.

Notably this failure occurs even if the `--force` flag is provided to
`pd migrate`, so I've updated the docstring on that flag accordingly.