UIP-6: App Version Safeguard (#4919)
## Describe your changes

This implements UIP 6, creating an "app version safeguard", to try and
prevent the wrong version of PD from being started against existing
state, or running the wrong version of PD.

This code should be immediately implementable as a non-breaking
*point-release*, which should *immediately* provide a safeguard against
forgetting to upgrade to the next major version of the software before
the next migration.
This should happen because nodes running the point release will start
writing, to non-consensus state, the app version they have. The current
migration in 0.80 will refuse to run unless this app version key is
empty, or exactly the *previous version*, preventing forgetting to
upgrade to the next version pre migration. Furthermore, the next
migration should do the same, but with the next app version, so that it
will not allow skipping the previous migration.

For testing, I think we should:
- test that we can take an existing node, and then start it with the new
code, stop it, and then start it again
- test that trying to run `pd migrate` _fails_ because of the newly
added safeguard.

## Issue ticket number and link

Closes #4793.

Implements https://github.com/penumbra-zone/UIPs/pull/10.

## Checklist before requesting a review

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> This code is sensitive to potentially breaking consensus, but the
testing plan above should confirm that it is safe for a point release.
If it is not safe for a point release, the code should be amended so
that it is.

---------

Co-authored-by: @erwanor <erwanor@users.noreply.github.com>
Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>