chore: update f4jumble to 0.1.0 (#4926)
## Describe your changes

Updates the old git dependency to a suitable published version from
crates.io. To review, see the diff review posted in #4921. I'd prefer
that reviewers replicate that work, and confirm that 1) the procedure
used for obtaining relevant sources was sound; and 2) the code changes
in the dep introduced by this diff are acceptable.

## Issue ticket number and link

Refs #4921.

## Checklist before requesting a review

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> The dep appears to function the same in pre and post version bump, as
does our modified error handling of its return values. Reviewers should
confirm those statements are true, and if so, then there's no affect on
consensus.