fix: update dependencies in parameter-setup crate (#4930)
## Describe your changes

The parameter-setup crate was broken due to outdated dependencies. The
options here are:
1. add this tool to the workspace
2. keep the dependencies updated instead of using workspace inheritance

Option 1 is more maintainable. Even though the parameters are fixed now
due to mainnet, folks may try to use this crate for debugging - the
scenario that motivated this commit.

To test: I followed the steps in
https://guide.penumbra.zone/dev/parameter_setup to ensure we could run
the parameter setup.

## Issue ticket number and link

This was discovered while debugging #4927

## Checklist before requesting a review

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > this is a dev only change