pindexer: insights: correct calculation of total value shielded (#4932)
## Describe your changes

This corrects the calculation of the total value shielded. Previously,
were were adding the absolute value of outbound transfers, as well as
refunds to the total. Instead, we just need to consider deposits.

To test, we should look at current mainnet data, and make sure that the
total for USDC is changing appropriately.

## Issue ticket number and link

Closes #4925

## Checklist before requesting a review

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > REPLACE THIS TEXT WITH RATIONALE (CAN BE BRIEF)