ibc: handle height in gRPC request metadata if specified (#4905)
## Describe your changes

Check if height is specified in the metadata for relevant IBC gRPC
queries and get the respective snapshot if it is provided. If the height
is not specified in the metadata, the latest snapshot is used, which is
the existing behaviour.

Reasoning: since astria uses only gRPC queries for all the IBC queries
in the hermes impl, but the gRPC proto messages don't contain the query
height, we implemented a fix by putting the height in the metadata.
Without this fix, our hermes impl would submit unexpected proofs on
occasion, as the proofs would be off-by-one height than that was
expected.

I opened a PR to update the protos, but the response by the maintainer
of ibc-go suggests to me that they intend for having the height in the
header be the actual fix (https://github.com/cosmos/ibc-go/pull/7303).

## Issue ticket number and link

n/a

## Checklist before requesting a review

- [ ] I have added guiding text to explain how a reviewer should test
these changes.

- [ ] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> only affects the IBC gRPC server, and does not affect existing
behaviour

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>