UIP-8: Transparent Addresses (#4950)
This commit experiments with a "truncated" address format, which aims at
maximizing compatibility with other chains. This is kind of a janky
hack, but may be worth doing anyways -- and could allow users to work
around current and future Bech32m encoding compatibility issues (though
our goal should be to ensure compatibility).

- Uses Bech32, like other Cosmos chains
- Fits in 32 bytes -- looks exactly like a cosmos except for a different
Bech32 HRP
- As an alternate encoding, requires no changes to other parts of
Penumbra, can be parsed directly into an `Address`

- A third address format adds UX complexity
- The truncated address points at a random account number
- Only one truncated address per IVK (reuse of the address causes
linkability)
- Not compatible with FMD (clues created with the truncated address are
not detectable)

A Penumbra address has three components: the diversifier `d`, the
transmission key `pk_d`, and the clue key `ck_d`. To truncate the
address to a 32-byte encoding:
- The diversifier is required to be `[0; 16]`, omitted from the
encoding, and filled in with the hardcoded value during decoding;
- The transmission key is encoded as 32 bytes;
- The clue key is stripped, and filled in on the decoding side with the
identity element.

The fixed diversifier means that there's only one per IVK, and also that
the truncated address points at a random sub-account: the fixed value
`[0u8; 16]` is decrypted with the diversifier key to a random address
index. Stripping the clue key and filling it in with a dummy value means
that generated clues are not detectable, but should not have other
effects on the protocol, and since FMD is currently unused this seems
fine.

To allow the use of truncated addresses in IBC transfers, the
`Ics20Withdrawal` action would need to be changed to have a
`use_truncated_address` parameter.

No issue - starting point for forum discussion

- [ ] ~~I have added guiding text to explain how a reviewer should test
these changes.~~ N/A - starting point for discussion

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label.

---------

Co-authored-by: redshiftzero <jen@penumbralabs.xyz>