revert merge of 4966 to release/v80 (#4968)
## Describe your changes

The PR https://github.com/penumbra-zone/penumbra/pull/4966 was
mistakenly targeted against `release/v0.80.x`, which branch has fallen
out of sync with main. If we want to make a Penumbra point release, PRs
should target `main`, and we'll tag directly on main.

Eventually we may want to revive `release/v0.80.x` and use it for
backports, but that time is not now. By reverting the out-f-order merge
from https://github.com/penumbra-zone/penumbra/pull/4966, we leave the
door open to updating that branch in the future so it can be used to
manage backports.

## Issue ticket number and link

Just https://github.com/penumbra-zone/penumbra/pull/4966 is relevant.

## Checklist before requesting a review

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> this is a housekeeping commit, intended to cleanup git history and
avoid confusion.