build: publish workspace to crates.io
We want all the crates Penumbra protocol workspace to be published to
crates.io, so that external developers can depend on them in downstream
projects. For now, we'll skip doing so for the binaries, but we can
circle back on them.

The major obstacle is that the `penumbra-sdk-proof-params` create
contains binary keyfiles, which are managed in git-lfs. We cannot upload
the raw keyfiles to crates.io, because that'd result in a 100MB crate.
Instead, we use a custom script to revert the binary keyfiles to
plaintext lfs pointers immediately prior to publishing to crates.io,
which stays under the limit, and allows third-party tools to opt into
downloading the key material via the `download-proving-keys` feature.

The ratelimit of "20s" between crates publishing has been working
reliably when publishing alpha versions (i.e. an HTTP 429 is avoided).
Publishing all crates takes ~12m.

This commit also condenses several "alpha" series releases matching
`0.82.0-alpha.*`, culminating `1.0.0-alpha.0`. Most of these versions
have been published to crates.io, to evaluate CI behavior on publishing crates.

build: warn-only if download-proving-keys not set

The goal is to warn if keys are not set, but still permit the build.