Revert "penumbra: update ecosystem tendermint/ibc crates (#4980)"
This reverts commit 64c32ef4bf33c929560ae2fc46ac127030c01f68,
which constitutes the squash-merge of PR #4980. We're backing this
change out strictly to simplify release engineering: we want the `main`
branch to remain fully compatible with the `0.81.x` series,
and we'll continue QA of significant version changes in a parallel
release branch, `release/v0.82.x`. I'll handle preparing the latter
shortly.

See related discussion in #4988 & #4991.