Revert "penumbra: update ecosystem tendermint/ibc crates (#4980)" (#4993)
## Describe your changes
This reverts commit 64c32ef4bf33c929560ae2fc46ac127030c01f68, which
constitutes the squash-merge of PR #4980. We're backing this change out
strictly to simplify release engineering: we want the `main` branch to
remain fully compatible with the `0.81.x` series, and we'll continue QA
of significant version changes in a parallel release branch,
`release/v0.82.x`. I'll handle preparing the latter shortly.

## Issue ticket number and link

See related discussion in #4988 & #4991.

## Testing and review

This is a programmatic change, in that I simply ran `git revert
64c32ef4b`, wrote some notes into the commit message, and pushed it up.
I also made sure to rerun `just proto` to regenerate the protos, and
confirmed there are no changes. That's good, that's precisely what we
wanted to see.

Preferably this change would land before #4992, since #4992 changes
protos. I'll regenerate protos in 4992 on top of this once it lands on
main.

## Checklist before requesting a review

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> This commit is expressly intended to preserve protocol compatibility
with 0.81.x. Future work on QA to ensure compat with 0.82 will happen in
a separate branch.