protos(view): add a subaccount filter to the `OwnedPositionIds` rpc (#4992)
## Describe your changes

This PR resubmits the unmerged changes from
https://github.com/penumbra-zone/penumbra/pull/4985, targeting `main`
branch. Below is the original submission text for 4985, which is still
accurate here.

This PR:
- adds an `AddressIndex` field to the `OwnedPositionIds` rpc request and
response
- regenerate the codegen protobufs

It doesn't seem important to do the extra work to support filtering in
the rust view server. The natural consumers of that information are web
interfaces. Market makers using the rust implementation would probably
simply integrate it in their own OMS rather than relying on the view
service RPC.

One question I have is whether the buf publish action will be able to
pickup the change to the protos.

## Checklist before requesting a review

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > view service changes

---------

Signed-off-by: Erwan Or <erwan.ounn.84@gmail.com>
Co-authored-by: Erwan Or <erwanor@penumbralabs.xyz>
Co-authored-by: Erwan Or <erwan.ounn.84@gmail.com>
Co-authored-by: Lúcás Meier <lucas@cronokirby.com>