penumbra: update ecosystem tendermint/ibc crates (#4997)
## Describe your changes

Once again, we update significant dependencies on the main branch.
Significantly, we've removed all git-dependencies from the workspace,
which required upgrading `tonic` to upstream, which cascaded into other
dependencies updates. With this change, we make the workspace nearly
ready for publication to crates.io (#4978): still to come is a
superficial refactor to rename the crates to `penumbra-sdk-*`, which
I'll submit in a follow-up PR, to aid in review.

This reverts commit a4548708d17c520bc844e763fd68c18c2ab8920c.

## Issue ticket number and link


We've added and then removed this changeset a few times over the past
week:

* https://github.com/penumbra-zone/penumbra/pull/4973
* https://github.com/penumbra-zone/penumbra/pull/4980
* https://github.com/penumbra-zone/penumbra/pull/4993

Now that we've tested it extensively, it's time to merge it into main
and keep moving forward. Relevant version bumps were tracked in the
following issues:

* tonic #4400
* ibc-types #4682
* cnidarium #4956
* tendermint-rs (no issue)


## Testing and review
In order to be certain that the changes honor protocol compatibility, I
made sure to test syncing a mainnet fullnode based on this changeset
from height 2622918—which is shortly after the change upgrade to
0.81.0—to 3136597, which is current height at time of writing. This
gives us great confidence that the hard work toward ensuring that
changes like #4682 were achieved in a compatible way.

## Checklist before requesting a review

- [x] I have added guiding text to explain how a reviewer should test
these changes.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> see testing notes above: we're confident we've done the needful with
this patch

Co-authored-by: Erwan Or <erwanor@penumbralabs.xyz>
Co-authored-by: Erwan Or <erwan.ounn.84@gmail.com>
Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>